feat: 예외 처리를 위한 Controller를 따로 생성한다.
